// src/app/shared/interfaces/adm-financeiras/dashboards/dashboard.dto.ts
export interface SimpleDashboardDto {
    assinatura_venda_valor: number;
    assinatura_venda_quantidade: number;

    produto_venda_valor: number;
    produto_venda_quantidade: number;

    servico_prestado_valor: number;
    servico_prestado_quantidade: number;

    cash_in_valor: number;
    cash_in_quantidade: number;

    cash_out_valor: number;
    cash_out_quantidade: number;

    cash_total_valor: number;
}